refactor(arrow-icons): update VerticalList and ImageGallery
