+++ packages/components/src/ArrowIcon/styles.css
@@ -14,5 +14,4 @@
   right: 0;
   z-index: 1;
   height: 100%;
-
 }
